rename liquidity_pools.rs to foreign_investments.rs
