+++ runtime/integration-tests/src/cases.rs
@@ -4,8 +4,8 @@ mod block_rewards;
 mod currency_conversions;
 mod ethereum_transaction;
 mod example;
+mod foreign_investments;
 mod investments;
-mod liquidity_pools;
 mod loans;
 mod lp;
 mod oracles;
